+++ src/assign/propagate.rs
@@ -280,6 +280,7 @@ impl PropagateIF for AssignStack {
             self.num_restart += 1;
             self.cpr_ema.update(self.num_conflict);
         } else {
+            #[cfg(feature = "assign_rate")]
             self.assign_rate.update(
                 self.num_vars
                     - self.num_asserted_vars