+++ src/state.rs
@@ -1,8 +1,8 @@
 /// Crate `state` is a collection of internal data.
 use {
     crate::{
-        assign, cdb, processor,
-        solver::{self, SolverEvent, StageManager},
+        assign, cdb,
+        solver::{RestartManager, SolverEvent, StageManager},
         types::*,
     },
     std::{
@@ -12,7 +12,7 @@ use {
         time::{Duration, Instant},
     },
 };
-const PROGRESS_REPORT_ROWS: usize = 8;
+const PROGRESS_REPORT_ROWS: usize = 7;
 
 /// API for state/statistics management, providing [`progress`](`crate::state::StateIF::progress`).
 pub trait StateIF {
@@ -24,31 +24,33 @@ pub trait StateIF {
     /// write a header of stat data to stdio.
     fn progress_header(&mut self);
     /// write stat data to stdio.
-    fn progress<A, C, E, R>(&mut self, asg: &A, cdb: &C, elim: &E, rst: &R)
+    fn progress<A, C>(&mut self, asg: &A, cdb: &C)
     where
         A: PropertyDereference<assign::property::Tusize, usize>
             + PropertyReference<assign::property::TEma, EmaView>,
         C: PropertyDereference<cdb::property::Tusize, usize>
             + PropertyDereference<cdb::property::Tf64, f64>
-            + PropertyReference<cdb::property::TEma, EmaView>,
-        E: PropertyDereference<processor::property::Tusize, usize>,
-        R: PropertyDereference<solver::restart::property::Tusize, usize>
-            + PropertyDereference<solver::restart::property::Tf64, f64>;
+            + PropertyReference<cdb::property::TEma, EmaView>;
     /// write a short message to stdout.
     fn flush<S: AsRef<str>>(&self, mes: S);
     /// write a one-line message as log.
-    fn log<S: AsRef<str>>(&mut self, tick: usize, mes: S);
+    fn log<S: AsRef<str>>(&mut self, tick: Option<(Option<usize>, Option<usize>, usize)>, mes: S);
 }
 
 /// stat index.
 #[derive(Clone, Eq, PartialEq)]
 pub enum Stat {
+    Restart,
     /// the number of vivification
     Vivification,
     /// the number of vivified (shrunk) clauses
     VivifiedClause,
     /// the number of vivified (asserted) vars
     VivifiedVar,
+    /// the number of invocations of simplify
+    Simplify,
+    /// the number of subsumed clause by processor
+    SubsumedClause,
     /// don't use this dummy (sentinel at the tail).
     EndOfStatIndex,
 }
@@ -83,8 +85,12 @@ impl IndexMut<Stat> for [usize] {
 pub struct State {
     /// solver configuration
     pub config: Config,
+    /// the problem.
+    pub cnf: CNFDescription,
     /// collection of statistics data
     pub stats: [usize; Stat::EndOfStatIndex as usize],
+    // Restart
+    pub restart: RestartManager,
     /// StageManager
     pub stm: StageManager,
     /// problem description
@@ -128,7 +134,9 @@ impl Default for State {
     fn default() -> State {
         State {
             config: Config::default(),
+            cnf: CNFDescription::default(),
             stats: [0; Stat::EndOfStatIndex as usize],
+            restart: RestartManager::default(),
             stm: StageManager::default(),
             target: CNFDescription::default(),
             reflection_interval: 10_000,
@@ -180,6 +188,8 @@ impl Instantiate for State {
     fn instantiate(config: &Config, cnf: &CNFDescription) -> State {
         State {
             config: config.clone(),
+            cnf: cnf.clone(),
+            restart: RestartManager::instantiate(config, cnf),
             stm: StageManager::instantiate(config, cnf),
             target: cnf.clone(),
             time_limit: config.c_timeout,
@@ -196,7 +206,10 @@ impl Instantiate for State {
             SolverEvent::Eliminate(_) => (),
             SolverEvent::Instantiate => (),
             SolverEvent::Reinitialize => (),
-            SolverEvent::Restart => (),
+            SolverEvent::Restart => {
+                self[Stat::Restart] += 1;
+                self.restart.handle(SolverEvent::Restart);
+            }
             SolverEvent::Stage(_) => (),
 
             #[cfg(feature = "clause_vivification")]
@@ -330,28 +343,39 @@ impl StateIF for State {
             stdout().flush().unwrap();
         }
     }
-    fn log<S: AsRef<str>>(&mut self, tick: usize, mes: S) {
+    fn log<S: AsRef<str>>(&mut self, tick: Option<(Option<usize>, Option<usize>, usize)>, mes: S) {
         if self.config.splr_interface && !self.config.quiet_mode && !self.config.use_log {
-            self.log_messages
-                .insert(0, format!("[{:>10}] {}", tick, mes.as_ref()));
+            self.log_messages.insert(
+                0,
+                match tick {
+                    Some((Some(seg), Some(cyc), stg)) => {
+                        format!("stage({:>2},{:>4},{:>5}): {}", seg, cyc, stg, mes.as_ref())
+                    }
+                    Some((None, Some(cyc), stg)) => {
+                        format!("stage(  ,{:>4},{:>5}): {}", cyc, stg, mes.as_ref())
+                    }
+                    Some((None, None, stg)) => {
+                        format!("stage(  ,    ,{:>5}): {}", stg, mes.as_ref())
+                    }
+                    Some(_) => unreachable!(),
+                    None => format!("### {}", mes.as_ref()),
+                },
+            );
         }
     }
     /// `mes` should be shorter than or equal to 9, or 8 + a delimiter.
     #[allow(clippy::cognitive_complexity)]
-    fn progress<A, C, E, R>(&mut self, asg: &A, cdb: &C, elim: &E, rst: &R)
+    fn progress<A, C>(&mut self, asg: &A, cdb: &C)
     where
         A: PropertyDereference<assign::property::Tusize, usize>
             + PropertyReference<assign::property::TEma, EmaView>,
         C: PropertyDereference<cdb::property::Tusize, usize>
             + PropertyDereference<cdb::property::Tf64, f64>
             + PropertyReference<cdb::property::TEma, EmaView>,
-        E: PropertyDereference<processor::property::Tusize, usize>,
-        R: PropertyDereference<solver::restart::property::Tusize, usize>
-            + PropertyDereference<solver::restart::property::Tf64, f64>,
     {
         if !self.config.splr_interface || self.config.quiet_mode {
             self.log_messages.clear();
-            self.record_stats(asg, cdb, elim, rst);
+            self.record_stats(asg, cdb);
             return;
         }
 
@@ -376,20 +400,15 @@ impl StateIF for State {
         let cdb_num_bi_clause = cdb.derefer(cdb::property::Tusize::NumBiClause);
         let cdb_num_lbd2 = cdb.derefer(cdb::property::Tusize::NumLBD2);
         let cdb_num_learnt = cdb.derefer(cdb::property::Tusize::NumLearnt);
-        let cdb_lbd_of_dp: f64 = cdb.derefer(cdb::property::Tf64::LiteralBlockEntanglement);
-
-        let elim_num_full = elim.derefer(processor::property::Tusize::NumFullElimination);
-        let elim_num_sub = elim.derefer(processor::property::Tusize::NumSubsumedClause);
-
-        let rst_num_blk: usize = rst.derefer(solver::restart::property::Tusize::NumBlock);
-        let rst_num_rst: usize = rst.derefer(solver::restart::property::Tusize::NumRestart);
+        let cdb_lb_ent: f64 = cdb.derefer(cdb::property::Tf64::LiteralBlockEntanglement);
+        let rst_num_rst: usize = self[Stat::Restart];
         let rst_asg: &EmaView = asg.refer(assign::property::TEma::AssignRate);
         let rst_lbd: &EmaView = cdb.refer(cdb::property::TEma::LBD);
-        let rst_lbd_thr: f64 = rst.derefer(solver::restart::property::Tf64::RestartThreshold);
+        let rst_eng: f64 = self.restart.penetration_energy_charged;
         let stg_segment: usize = self.stm.current_segment();
 
         if self.config.use_log {
-            self.dump(asg, cdb, rst);
+            self.dump(asg, cdb);
             return;
         }
         self.progress_cnt += 1;
@@ -460,23 +479,31 @@ impl StateIF for State {
                 cdb_num_clause - cdb_num_learnt
             ),
         );
+        self[LogUsizeId::StageSegment] = stg_segment;
+        self[LogF64Id::RestartEnergy] = rst_eng;
+        self[LogF64Id::TrendASG] = rst_asg.trend();
         println!(
-            "\x1B[2K     Restart|#BLK:{}, #RST:{}, thrs:{}, #seg:{}",
-            im!("{:>9}", self, LogUsizeId::RestartBlock, rst_num_blk),
-            im!("{:>9}", self, LogUsizeId::Restart, rst_num_rst),
-            fm!("{:>9.4}", self, LogF64Id::RestartThreshold, rst_lbd_thr),
-            im!("{:>9}", self, LogUsizeId::StageSegment, stg_segment),
-        );
-        println!(
-            "\x1B[2K         LBD|trnd:{}, avrg:{}, entg:{}, /dpc:{}",
-            fm!("{:>9.4}", self, LogF64Id::TrendLBD, rst_lbd.trend()),
-            fm!("{:>9.4}", self, LogF64Id::EmaLBD, rst_lbd.get_fast()),
+            "\x1B[2K    Conflict|entg:{}, cLvl:{}, bLvl:{}, /cpr:{}",
             fm!(
                 "{:>9.4}",
                 self,
                 LogF64Id::LiteralBlockEntanglement,
-                cdb_lbd_of_dp
+                cdb_lb_ent
             ),
+            fm!("{:>9.4}", self, LogF64Id::CLevel, self.c_lvl.get()),
+            fm!("{:>9.4}", self, LogF64Id::BLevel, self.b_lvl.get()),
+            fm!(
+                "{:>9.2}",
+                self,
+                LogF64Id::ConflictPerRestart,
+                asg_cpr_ema.get()
+            )
+        );
+        println!(
+            "\x1B[2K    Learning|avrg:{}, trnd:{}, #RST:{}, /dpc:{}",
+            fm!("{:>9.4}", self, LogF64Id::EmaLBD, rst_lbd.get_fast()),
+            fm!("{:>9.4}", self, LogF64Id::TrendLBD, rst_lbd.trend()),
+            im!("{:>9}", self, LogUsizeId::Restart, rst_num_rst),
             fm!(
                 "{:>9.2}",
                 self,
@@ -485,26 +512,19 @@ impl StateIF for State {
             ),
         );
         println!(
-            "\x1B[2K    Conflict|tASG:{}, cLvl:{}, bLvl:{}, /ppc:{}",
-            fm!("{:>9.4}", self, LogF64Id::TrendASG, rst_asg.trend()),
-            fm!("{:>9.2}", self, LogF64Id::CLevel, self.c_lvl.get()),
-            fm!("{:>9.2}", self, LogF64Id::BLevel, self.b_lvl.get()),
-            fm!(
-                "{:>9.2}",
-                self,
-                LogF64Id::PropagationPerConflict,
-                asg_ppc_ema.get()
-            ),
-        );
-        println!(
-            "\x1B[2K        misc|vivC:{}, subC:{}, core:{}, /cpr:{}",
+            "\x1B[2K        misc|vivC:{}, subC:{}, core:{}, /ppc:{}",
             im!(
                 "{:>9}",
                 self,
                 LogUsizeId::VivifiedClause,
                 self[Stat::VivifiedClause]
             ),
-            im!("{:>9}", self, LogUsizeId::SubsumedClause, elim_num_sub),
+            im!(
+                "{:>9}",
+                self,
+                LogUsizeId::SubsumedClause,
+                self[Stat::SubsumedClause]
+            ),
             im!(
                 "{:>9}",
                 self,
@@ -518,11 +538,10 @@ impl StateIF for State {
             fm!(
                 "{:>9.2}",
                 self,
-                LogF64Id::ConflictPerRestart,
-                asg_cpr_ema.get()
-            )
+                LogF64Id::PropagationPerConflict,
+                asg_ppc_ema.get()
+            ),
         );
-        self[LogUsizeId::Simplify] = elim_num_full;
         self[LogUsizeId::Stage] = self.stm.current_stage();
         self[LogUsizeId::StageCycle] = self.stm.current_cycle();
         self[LogUsizeId::Vivify] = self[Stat::Vivification];
@@ -532,16 +551,13 @@ impl StateIF for State {
 
 impl State {
     #[allow(clippy::cognitive_complexity)]
-    fn record_stats<A, C, E, R>(&mut self, asg: &A, cdb: &C, elim: &E, rst: &R)
+    fn record_stats<A, C>(&mut self, asg: &A, cdb: &C)
     where
         A: PropertyDereference<assign::property::Tusize, usize>
             + PropertyReference<assign::property::TEma, EmaView>,
         C: PropertyDereference<cdb::property::Tusize, usize>
             + PropertyDereference<cdb::property::Tf64, f64>
             + PropertyReference<cdb::property::TEma, EmaView>,
-        E: PropertyDereference<processor::property::Tusize, usize>,
-        R: PropertyDereference<solver::restart::property::Tusize, usize>
-            + PropertyDereference<solver::restart::property::Tf64, f64>,
     {
         self[LogUsizeId::NumConflict] = asg.derefer(assign::property::Tusize::NumConflict);
         self[LogUsizeId::NumDecision] = asg.derefer(assign::property::Tusize::NumDecision);
@@ -558,15 +574,12 @@ impl State {
         self[LogUsizeId::BiClause] = cdb.derefer(cdb::property::Tusize::NumBiClause);
         self[LogUsizeId::PermanentClause] =
             cdb.derefer(cdb::property::Tusize::NumClause) - self[LogUsizeId::RemovableClause];
-        self[LogUsizeId::RestartBlock] = rst.derefer(solver::restart::property::Tusize::NumBlock);
-        self[LogUsizeId::Restart] = rst.derefer(solver::restart::property::Tusize::NumRestart);
+        self[LogUsizeId::Restart] = self[Stat::Restart];
         self[LogUsizeId::Stage] = self.stm.current_stage();
         self[LogUsizeId::StageCycle] = self.stm.current_cycle();
         self[LogUsizeId::StageSegment] = self.stm.max_scale();
-        self[LogUsizeId::Simplify] = elim.derefer(processor::property::Tusize::NumFullElimination);
-
-        self[LogUsizeId::SubsumedClause] =
-            elim.derefer(processor::property::Tusize::NumSubsumedClause);
+        self[LogUsizeId::Simplify] = self[Stat::Simplify];
+        self[LogUsizeId::SubsumedClause] = self[Stat::SubsumedClause];
         self[LogUsizeId::VivifiedClause] = self[Stat::VivifiedClause];
         self[LogUsizeId::VivifiedVar] = self[Stat::VivifiedVar];
         self[LogUsizeId::Vivify] = self[Stat::Vivification];
@@ -700,11 +713,10 @@ impl State {
              c ========================================================================================================="
         );
     }
-    fn dump<A, C, R>(&mut self, asg: &A, cdb: &C, rst: &R)
+    fn dump<A, C>(&mut self, asg: &A, cdb: &C)
     where
         A: PropertyDereference<assign::property::Tusize, usize>,
         C: PropertyDereference<cdb::property::Tusize, usize>,
-        R: PropertyDereference<solver::restart::property::Tusize, usize>,
     {
         self.progress_cnt += 1;
         let asg_num_vars = asg.derefer(assign::property::Tusize::NumVar);
@@ -718,11 +730,9 @@ impl State {
         let cdb_num_lbd2 = cdb.derefer(cdb::property::Tusize::NumLBD2);
         let cdb_num_learnt = cdb.derefer(cdb::property::Tusize::NumLearnt);
         let cdb_num_reduction = cdb.derefer(cdb::property::Tusize::NumReduction);
-        let rst_num_block = rst.derefer(solver::restart::property::Tusize::NumBlock);
         println!(
-            "c | {:>8}  {:>8} {:>8} | {:>7} {:>8} {:>8} |  {:>4}  {:>8} {:>7} {:>8} | {:>6.3} % |",
+            "c | {:>8} {:>8} | {:>7} {:>8} {:>8} |  {:>4}  {:>8} {:>7} {:>8} | {:>6.3} % |",
             asg_num_restart,                           // restart
-            rst_num_block,                             // blocked
             asg_num_conflict / asg_num_restart.max(1), // average cfc (Conflict / Restart)
             asg_num_unasserted_vars,                   // alive vars
             cdb_num_clause - cdb_num_learnt,           // given clauses
@@ -735,13 +745,12 @@ impl State {
         );
     }
     #[allow(dead_code)]
-    fn dump_details<'r, A, C, E, R, V>(&mut self, asg: &A, cdb: &C, rst: &'r R)
+    fn dump_details<A, C>(&mut self, asg: &A, cdb: &C)
     where
         A: PropertyDereference<assign::property::Tusize, usize>
             + PropertyReference<assign::property::TEma, EmaView>,
         C: PropertyDereference<cdb::property::Tusize, usize>
             + PropertyReference<cdb::property::TEma, EmaView>,
-        R: PropertyDereference<solver::restart::property::Tusize, usize>,
     {
         self.progress_cnt += 1;
         let asg_num_vars = asg.derefer(assign::property::Tusize::NumVar);
@@ -752,7 +761,6 @@ impl State {
         let asg_num_restart = asg.derefer(assign::property::Tusize::NumRestart);
         let cdb_num_clause = cdb.derefer(cdb::property::Tusize::NumClause);
         let cdb_num_learnt = cdb.derefer(cdb::property::Tusize::NumLearnt);
-        let rst_num_block = rst.derefer(solver::restart::property::Tusize::NumBlock);
         let rst_asg = asg.refer(assign::property::TEma::AssignRate);
         let rst_lbd = cdb.refer(cdb::property::TEma::LBD);
 
@@ -768,7 +776,7 @@ impl State {
             cdb_num_learnt,
             cdb_num_clause,
             0,
-            rst_num_block,
+            0,
             asg_num_restart,
             rst_asg.trend(),
             rst_lbd.get(),
@@ -811,8 +819,6 @@ pub enum LogUsizeId {
     //## restart
     //
     Restart,
-    RestartBlock,
-    RestartCancel,
 
     //
     //## stage
@@ -850,7 +856,7 @@ pub enum LogF64Id {
     ConflictPerRestart,
     PropagationPerConflict,
     LiteralBlockEntanglement,
-    RestartThreshold,
+    RestartEnergy,
     End,
 }